+++ Atarashii/src/net/somethingdreadful/MAL/api/response/AnimeManga/Manga.java
@@ -1,10 +1,12 @@
-package net.somethingdreadful.MAL.api.response;
+package net.somethingdreadful.MAL.api.response.AnimeManga;
 
 import android.database.Cursor;
 
 import com.google.gson.Gson;
 import com.google.gson.annotations.SerializedName;
 
+import net.somethingdreadful.MAL.api.response.GenericRecord;
+import net.somethingdreadful.MAL.api.response.RecordStub;
 import net.somethingdreadful.MAL.sql.MALSqlHelper;
 
 import java.io.Serializable;
