+++ Atarashii/src/net/somethingdreadful/MAL/dialog/StatusPickerDialogFragment.java
@@ -11,9 +11,9 @@ import android.widget.RadioGroup.OnCheckedChangeListener;
 import net.somethingdreadful.MAL.DetailView;
 import net.somethingdreadful.MAL.R;
 import net.somethingdreadful.MAL.api.MALApi.ListType;
-import net.somethingdreadful.MAL.api.response.Anime;
+import net.somethingdreadful.MAL.api.response.AnimeManga.Anime;
+import net.somethingdreadful.MAL.api.response.AnimeManga.Manga;
 import net.somethingdreadful.MAL.api.response.GenericRecord;
-import net.somethingdreadful.MAL.api.response.Manga;
 
 public class StatusPickerDialogFragment extends DialogFragment implements OnCheckedChangeListener {
