+++ Atarashii/src/net/somethingdreadful/MAL/forum/HtmlUtil.java
@@ -6,10 +6,10 @@ import net.somethingdreadful.MAL.DateTools;
 import net.somethingdreadful.MAL.ProfileActivity;
 import net.somethingdreadful.MAL.R;
 import net.somethingdreadful.MAL.Theme;
-import net.somethingdreadful.MAL.api.response.Activity;
 import net.somethingdreadful.MAL.api.response.Forum;
 import net.somethingdreadful.MAL.api.response.ForumMain;
-import net.somethingdreadful.MAL.api.response.User;
+import net.somethingdreadful.MAL.api.response.UserProfile.Activity;
+import net.somethingdreadful.MAL.api.response.UserProfile.User;
 
 import java.io.IOException;
 import java.io.InputStream;
