+++ Atarashii/src/net/somethingdreadful/MAL/profile/ProfileDetailsMAL.java
@@ -31,7 +31,7 @@ import net.somethingdreadful.MAL.ProfileActivity;
 import net.somethingdreadful.MAL.R;
 import net.somethingdreadful.MAL.Theme;
 import net.somethingdreadful.MAL.api.MALApi;
-import net.somethingdreadful.MAL.api.response.User;
+import net.somethingdreadful.MAL.api.response.UserProfile.User;
 
 import org.apache.commons.lang3.text.WordUtils;
