+++ Atarashii/src/net/somethingdreadful/MAL/tasks/FriendsNetworkTaskFinishedListener.java
@@ -1,9 +1,9 @@
 package net.somethingdreadful.MAL.tasks;
 
-import net.somethingdreadful.MAL.api.response.User;
+import net.somethingdreadful.MAL.api.response.UserProfile.User;
 
 import java.util.ArrayList;
 
 public interface FriendsNetworkTaskFinishedListener {
-    public void onFriendsNetworkTaskFinished(ArrayList<User> result);
+    void onFriendsNetworkTaskFinished(ArrayList<User> result);
 }
