+++ Atarashii/src/net/somethingdreadful/MAL/tasks/UserNetworkTask.java
@@ -9,7 +9,7 @@ import com.crashlytics.android.Crashlytics;
 import net.somethingdreadful.MAL.MALManager;
 import net.somethingdreadful.MAL.account.AccountService;
 import net.somethingdreadful.MAL.api.MALApi;
-import net.somethingdreadful.MAL.api.response.User;
+import net.somethingdreadful.MAL.api.response.UserProfile.User;
 
 public class UserNetworkTask extends AsyncTask<String, Void, User> {
     Context context;
