+++ Atarashii/src/net/somethingdreadful/MAL/tasks/WriteDetailTask.java
@@ -13,9 +13,9 @@ import net.somethingdreadful.MAL.RecordStatusUpdatedReceiver;
 import net.somethingdreadful.MAL.account.AccountService;
 import net.somethingdreadful.MAL.api.MALApi;
 import net.somethingdreadful.MAL.api.MALApi.ListType;
-import net.somethingdreadful.MAL.api.response.Anime;
+import net.somethingdreadful.MAL.api.response.AnimeManga.Anime;
+import net.somethingdreadful.MAL.api.response.AnimeManga.Manga;
 import net.somethingdreadful.MAL.api.response.GenericRecord;
-import net.somethingdreadful.MAL.api.response.Manga;
 
 import retrofit.RetrofitError;
